Migrate F-5E-3 to new module type
