+++ Scripts/DCS-BIOS/doc/Addresses.h
@@ -13231,24 +13231,20 @@
 #define F_5E_3_HSI_CRS_A 0x76D0
 #define F_5E_3_HSI_CRS_DEV 0x76D8, 0xFFFF, 0
 #define F_5E_3_HSI_CRS_DEV_A 0x76D8
-#define F_5E_3_HSI_CRS_KNOB 0x76CE, 0xFFFF, 0
-#define F_5E_3_HSI_CRS_KNOB_A 0x76CE
 #define F_5E_3_HSI_DEV 0x76E8, 0xFFFF, 0
 #define F_5E_3_HSI_DEV_A 0x76E8
 #define F_5E_3_HSI_HDG 0x76D2, 0xFFFF, 0
 #define F_5E_3_HSI_HDG_A 0x76D2
-#define F_5E_3_HSI_HDG_KNOB 0x76CC, 0xFFFF, 0
-#define F_5E_3_HSI_HDG_KNOB_A 0x76CC
 #define F_5E_3_HSI_OFF 0x76E6, 0xFFFF, 0
 #define F_5E_3_HSI_OFF_A 0x76E6
-#define F_5E_3_HSI_Range_1 0x76DE, 0xFFFF, 0
-#define F_5E_3_HSI_Range_10 0x76DC, 0xFFFF, 0
-#define F_5E_3_HSI_Range_100 0x76DA, 0xFFFF, 0
-#define F_5E_3_HSI_Range_100_A 0x76DA
-#define F_5E_3_HSI_Range_10_A 0x76DC
-#define F_5E_3_HSI_Range_1_A 0x76DE
-#define F_5E_3_HSI_Range_FLAG 0x768E, 0x0010, 4
-#define F_5E_3_HSI_Range_FLAG_AM 0x768E, 0x0010
+#define F_5E_3_HSI_RANGE_1 0x76DE, 0xFFFF, 0
+#define F_5E_3_HSI_RANGE_10 0x76DC, 0xFFFF, 0
+#define F_5E_3_HSI_RANGE_100 0x76DA, 0xFFFF, 0
+#define F_5E_3_HSI_RANGE_100_A 0x76DA
+#define F_5E_3_HSI_RANGE_10_A 0x76DC
+#define F_5E_3_HSI_RANGE_1_A 0x76DE
+#define F_5E_3_HSI_RANGE_FLAG 0x768E, 0x0010, 4
+#define F_5E_3_HSI_RANGE_FLAG_AM 0x768E, 0x0010
 #define F_5E_3_HSI_TO_FROM 0x76E4, 0xFFFF, 0
 #define F_5E_3_HSI_TO_FROM_A 0x76E4
 #define F_5E_3_IAS_SET 0x76A6, 0xFFFF, 0
@@ -13263,10 +13259,16 @@
 #define F_5E_3_IFF4_M1 0x7688, 0x0600, 9
 #define F_5E_3_IFF4_M2 0x7688, 0x1800, 11
 #define F_5E_3_IFF4_M3 0x7688, 0x6000, 13
-#define F_5E_3_IFF4_MASTER 0x768C, 0xF000, 12
+#define F_5E_3_IFF4_MASTER 0x7688, 0x0070, 4
 #define F_5E_3_IFF4_MASTER_PULL 0x768E, 0x0002, 1
 #define F_5E_3_IFF4_MASTER_PULL_AM 0x768E, 0x0002
 #define F_5E_3_IFF4_MC 0x768A, 0x0003, 0
+#define F_5E_3_IFF4_MODE1_SELECTOR1 0x768A, 0x03C0, 6
+#define F_5E_3_IFF4_MODE1_SELECTOR2 0x768A, 0x3C00, 10
+#define F_5E_3_IFF4_MODE3_SELECTOR1 0x768C, 0x000F, 0
+#define F_5E_3_IFF4_MODE3_SELECTOR2 0x768C, 0x00F0, 4
+#define F_5E_3_IFF4_MODE3_SELECTOR3 0x768C, 0x0F00, 8
+#define F_5E_3_IFF4_MODE3_SELECTOR4 0x768C, 0xF000, 12
 #define F_5E_3_IFF4_MON_CONTR 0x7688, 0x0180, 7
 #define F_5E_3_IFF4_RAD 0x768A, 0x000C, 2
 #define F_5E_3_IFF4_RAD_TEST_L 0x768A, 0x8000, 15
